+++ input/elasticapm/internal/modeldecoder/v2/decoder.go
@@ -1120,7 +1120,7 @@ func mapToStracktraceModel(from []stacktraceFrame, out model.Stacktrace) {
 			fr.Classname = eventFrame.Classname.Val
 		}
 		if eventFrame.ColumnNumber.IsSet() {
-			val := eventFrame.ColumnNumber.Val
+			val := uint32(eventFrame.ColumnNumber.Val)
 			fr.Colno = &val
 		}
 		if eventFrame.ContextLine.IsSet() {
@@ -1137,7 +1137,7 @@ func mapToStracktraceModel(from []stacktraceFrame, out model.Stacktrace) {
 			fr.LibraryFrame = val
 		}
 		if eventFrame.LineNumber.IsSet() {
-			val := eventFrame.LineNumber.Val
+			val := uint32(eventFrame.LineNumber.Val)
 			fr.Lineno = &val
 		}
 		if eventFrame.Module.IsSet() {
